Move StreamZipper test helper to core
