+++ core/src/test/java/net/kemitix/naolo/core/StreamZipper.java
@@ -19,7 +19,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
 
-package net.kemitix.naolo.test;
+package net.kemitix.naolo.core;
 
 import java.util.List;
 import java.util.function.BiFunction;
